lighter verbalizer, without dependency on vocabulary
